[automessage] Xavier'fortune ;)
Let me not to the marriage of true minds
Admit impediments.  Love is not love
Which alters when it alteration finds,
Or bends with the remover to remove:
O, no! it is an ever-fixed mark,
That looks on tempests and is never shaken;
It is the star to every wandering bark,
Whose worth's unknown, although his height be taken.
Love's not Time's fool, though rosy lips and cheeks
Within his bending sickle's compass come;
Love alters not with his brief hours and weeks,
But bears it out even to the edge of doom.
If this be error and upon me proved,
I never writ, nor no man ever loved.

+++ trunk/AE-go_GameServer/src/com/aionemu/gameserver/network/aion/serverpackets/SM_PLAYER_INFO.java
@@ -121,7 +121,7 @@ public class SM_PLAYER_INFO extends AionServerPacket
 		int itemsCount = 0;
 		int row = 0;
 		while (totalEquipedItemsCount > 0) {
-			short slot = equipedItems.getEquipedItemSlotArray(row).getSlotMask();
+			int slot = equipedItems.getEquipedItemSlotArray(row).getSlotMask();
 			itemsCount = itemsCount + slot;
 			totalEquipedItemsCount = totalEquipedItemsCount-1;
 			row+=1;
